Add new qnnpack_add and qnnpack_maxpool op to C10 registry
Pull Request resolved: https://github.com/pytorch/pytorch/pull/24103

This change adds a quantized add and maxpool2d operation for pytorch mobile.

These operators follow the structure of qnnpack in terms of create/setup and run calls. The plan to refactor QNNPACK to make it more functional is currently for FC and Conv ops where the cost of create/setup is high.
For ops like add and maxpool the cost of calling create and setup in each operator invocation is negligible.

Once we migrate FC and Conv QNNPACK ops to be functional in nature, we will consider changing these ops as well to make it consistent.
ghstack-source-id: 88159376

Differential Revision: [D16734190](https://our.internmc.facebook.com/intern/diff/D16734190/)